RMA-Demo/services/upload-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
PyJWT==2.8.0
qrcode[pil]==7.4.2
Pillow==10.2.0
httpx==0.26.0
